etl/postgres_to_es/requirements.txt
flake8==4.0
python-dotenv==0.20.0
psycopg2-binary==2.9.4
pytest==6.2.5
flake8-bandit==2.1.2
bandit==1.7.2
requests==2.26.0
redis==4.3.1